- fixed random number issue

+++ DeenIslamLibrary/src/main/java/com/deenislam/sdk/viewmodels/UserTrackViewModel.kt
@@ -9,7 +9,7 @@ internal class UserTrackViewModel(
     private val repository: UserTrackRepository
 ) : ViewModel() {
 
-    fun trackUser(language:String,msisdn:String,pagename:String,trackingID:Long)
+    fun trackUser(language:String,msisdn:String,pagename:String,trackingID:Int)
     {
         viewModelScope.launch {
             repository.userTrack(
